Treat expression before passing to dataset endpoint
